feat(backend): support decimal quantities in venta items - BigDecimal instead of Integer

+++ src/main/java/com/merceria153/dto/ItemRequest.java
@@ -2,9 +2,10 @@ package com.merceria153.dto;
 
 import lombok.Data;
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Data
 public class ItemRequest {
     @NotNull private String productoId;
-    @NotNull private Integer cantidad;
+    @NotNull private BigDecimal cantidad;
 }
